Migrate to ES6 syntax
