+++ package.json
@@ -34,6 +34,7 @@
     "start": "rc-tools run server",
     "pub": "rc-tools run pub --babel-runtime",
     "lint": "rc-tools run lint",
+    "lint:fix": "rc-tools run lint --fix",
     "test": "jest",
     "prepublish": "rc-tools run guard",
     "coverage": "jest --coverage && cat ./coverage/lcov.info | coveralls"
@@ -67,7 +68,6 @@
     "babel-runtime": "^6.23.0",
     "classnames": "2.x",
     "component-classes": "1.x",
-    "create-react-class": "^15.5.2",
     "dom-scroll-into-view": "1.x",
     "prop-types": "^15.5.8",
     "rc-animate": "2.x",
